+++ libs/shared/testing/src/lib/fixtures/index.ts
@@ -1,2 +1,1 @@
 export * from './realms';
-export * from './users';